requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
datasets==2.19.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
executing==2.2.1
fastapi==0.110.0
fastjsonschema==2.21.2
filelock==3.20.3
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.61.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==1.3.7
idna==3.11
ipykernel==6.29.3
ipython==9.10.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-client==8.8.0
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.2
lark==1.3.1
loguru==0.7.2
markupsafe==3.0.3
matplotlib==3.8.3
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.16
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.5.3
notebook-shim==0.2.4
numpy==1.26.4
openai==1.54.0
packaging==26.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prometheus-client==0.24.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.5
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==4.0.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
qtconsole==5.7.0
qtpy==2.4.3
redis==5.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rouge-score==0.1.2
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
send2trash==2.1.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
starlette==0.36.3
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.2.1
torch-geometric==2.5.2
tornado==6.5.4
tqdm==4.66.2
traitlets==5.14.3
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
